support for both Pico and Pico Prism
